agregado maximizacion de ventana de ventas

+++ build/views/vendedor/facturacion.js
@@ -83,16 +83,17 @@ function getView(){
             return `
         <div class="row">
             <div id="panel-2" class="panel col-12">
+
                 <div class="panel-hdr">
                     <h2 id="txtTotalVenta" class="text-danger"></h2>
                     <div class="panel-toolbar">
-                        <button class="btn btn-panel" data-action="panel-collapse" data-toggle="tooltip" data-offset="0,10" data-original-title="Collapse"></button>
-                        
-                       
+                                               
+                        <button class="btn btn-warning" data-action="panel-fullscreen" data-toggle="tooltip" data-offset="0,10" data-original-title="Fullscreen">
+                            <i class="fal fa-angle-double-up"></i>
+                        </button>
                     </div>
                 </div>
                 <div class="panel-container show">
-
                     <div class="panel-content">
                         <div class="col-sm-12 col-md-8 col-lg-8 col-xl-8">
                             <div class="input-group">
@@ -111,7 +112,6 @@ function getView(){
                                 </div>
                             </div>
                         </div>
-
                         <div class="table-responsive">
                             <table class="table table-hover table-striped"><!--mt-5-->
                                 <thead>
@@ -126,9 +126,147 @@ function getView(){
                                 <tbody id="tblGridTempVentas"></tbody>
                             </table>
                         </div>
-                
                     </div>
                 </div>
+
+                <div id="containerModalesVentas"></div>
+
+
+            </div>
+        </div>  
+            `
+        },
+        gridTempVentaModalBusquedaProductos :()=>{
+            return `
+        <div class="row">
+            <div id="panel-2" class="panel col-12">
+
+
+                <div class="panel-hdr">
+                    <h2 id="txtTotalVenta" class="text-danger"></h2>
+                    <div class="panel-toolbar">
+                                               
+                        <button class="btn btn-warning" data-action="panel-fullscreen" data-toggle="tooltip" data-offset="0,10" data-original-title="Fullscreen">
+                            <i class="fal fa-angle-double-up"></i>
+                        </button>
+                    </div>
+                </div>
+                <div class="panel-container show">
+                    <div class="panel-content">
+                        <div class="col-sm-12 col-md-8 col-lg-8 col-xl-8">
+                            <div class="input-group">
+                                <select class="form-control col-3 shadow border-info" id="cmbTipoPrecio">
+                                    <option value="P">DET</option>
+                                    <option value="C">PreB</option>
+                                    <option value="B">PreC</option>
+                                    <option value="A">MAY</option>
+                                    <option value="K">CAMBIO</option>
+                                </select>
+                                <input id="txtBusqueda" type="text" ref="txtBusqueda" class="form-control col-7  shadow border-info" placeholder="Buscar código o descripción..." aria-label="" aria-describedby="button-addon4" />
+                                <div class="input-group-prepend">
+                                    <button class="btn btn-info waves-effect waves-themed shadow" type="button" id="btnBuscarProducto">
+                                        <i class="fal fa-search"></i>
+                                    </button>
+                                </div>
+                            </div>
+                        </div>
+                        <div class="table-responsive">
+                            <table class="table table-hover table-striped"><!--mt-5-->
+                                <thead>
+                                    <tr>
+                                        <th class="border-top-0 table-scale-border-bottom fw-700">Producto</th>
+                                        <th class="text-right border-top-0 table-scale-border-bottom fw-700">Medida</th>
+                                        <th class="text-center border-top-0 table-scale-border-bottom fw-700">Cant.</th>
+                                        <th class="text-right border-top-0 table-scale-border-bottom fw-700">Subtotal</th>
+                                        <th class="text-center border-top-0 table-scale-border-bottom fw-700"></th>
+                                    </tr>
+                                </thead>
+                                <tbody id="tblGridTempVentas"></tbody>
+                            </table>
+                        </div>
+                    </div>
+                </div>
+
+                <div class="modal fade" id="ModalBusqueda" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                    <div class="modal-dialog modal-lg modal-dialog-right" role="document">
+                        <div class="modal-content">
+                            <div class="modal-header">
+                                <label class="modal-title text-danger h3" id="">Resultados de la Búsqueda</label>
+                                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                    <span aria-hidden="true"><i class="fal fa-times"></i></span>
+                                </button>
+                            </div>
+                            <div class="modal-body">
+                            <table class="table table-responsive table-striped table-hover">
+                                <thead>
+                                    <tr>
+                                        <td>Producto</td>
+                                        <td>Precio</td>                         
+                                        <td></td>
+                                    </tr>
+                                </thead>
+                                <tbody id="tblResultadoBusqueda">
+                                
+                                </tbody>
+                            </table>
+                            </div>        
+                        </div>
+                    </div>
+                </div>
+
+                <div class="modal fade" id="ModalCantidadProducto" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                    <div class="modal-dialog modal-dialog-right" role="document">
+                        <div class="modal-content">
+                            <br><br><br><br><br>
+                            <div class="modal-header">
+                                <label class="modal-title" id="txtDesProducto">Azucar don Justo Cabal Kilo</label>
+                            </div>
+                            <div class="modal-body" align="right">
+                                <div class="col-8">
+                                    <div class="row">
+                                        <b id="txtCodMedida">UNIDAD</b>
+                                    </div>
+                                    <div class="form-group">
+                                        <div class="row">
+                                            <div class="input-group">  
+                                                <div class="input-group-prepend">
+                                                    <button class="btn btn-md btn-icon btn-round btn-info" id="btnCantidadDown">
+                                                        -
+                                                    </button>
+                                                </div>
+                                    
+                                                <input type="number" class="text-center form-control" id="txtCantidad" value="1">    
+                                    
+                                                <div class="input-group-append">
+                                                    <button class="btn btn-md btn-icon btn-round btn-info" id="btnCantidadUp">
+                                                        +
+                                                    </button>    
+                                                </div>
+                                            </div>                            
+                                        </div>                              
+                                    </div>
+                                    <div class="col-12">
+                                        <label>Precio: </label>
+                                        <label class="text-success" id="txtPrecioProducto">Q500</label>
+                                        <br>
+                                        <label>Subtotal:</label>
+                                        <label class="text-danger" id="txtSubTotal">Q500</label>
+                                    </div>
+                                    <br>
+                                    <div class="">
+                                        <button type="button" class="btn btn-outline-secondary btn-round" data-dismiss="modal" id="btnCancelarModalProducto">
+                                            <i class="fal fa-ban"></i>Cancelar
+                                        </button>
+                                        <button type="button" class="btn btn-primary btn-round" data-dismiss="modal" id="btnAgregarProducto">
+                                            <i class="fal fa-check"></i>Agregar
+                                        </button>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+
             </div>
         </div>  
             `
@@ -610,17 +748,15 @@ function getView(){
                                     <label>Subtotal:</label>
                                     <label class="text-danger" id="txtSubTotal">Q500</label>
                                 </div>
-
                                 <br>
                                 <div class="">
-                                    <button type="button" class="btn btn-secondary btn-round" data-dismiss="modal" id="btnCancelarModalProducto">
-                                        <i class="now-ui-icons ui-1_simple-remove"></i>Cancelar
+                                    <button type="button" class="btn btn-outline-secondary btn-round" data-dismiss="modal" id="btnCancelarModalProducto">
+                                        <i class="fal fa-ban"></i>Cancelar
                                     </button>
                                     <button type="button" class="btn btn-primary btn-round" data-dismiss="modal" id="btnAgregarProducto">
-                                        <i class="now-ui-icons ui-1_simple-add"></i>Agregar
+                                        <i class="fal fa-check"></i>Agregar
                                     </button>
                                 </div>
-
                             </div>
                         </div>
 
@@ -635,12 +771,18 @@ function getView(){
     root.innerHTML = view.encabezadoClienteDocumento() 
                 + view.gridTempVenta() 
                 + view.btnCobrar() 
-                + view.modalBusquedaProductos() + view.modalCantidadProducto()
+                //+ view.modalBusquedaProductos() 
+                //+ view.modalCantidadProducto()
                 + view.modalBusquedaCliente() 
                 + view.modalCobro() 
                 + view.modalNuevoCliente() 
-                + view.modalTerminar() 
-                + view.modalCantidadCalculadora();
+                + view.modalTerminar(); 
+                //+ view.modalCantidadCalculadora();
+
+    let containerModalesVentas = document.getElementById('containerModalesVentas');
+    containerModalesVentas.innerHTML = view.modalBusquedaProductos() 
+                                        + view.modalCantidadProducto()
+                                        + view.modalCantidadCalculadora();
 
 };
 
@@ -938,8 +1080,9 @@ async function fcnBusquedaProducto(idFiltro,idTablaResultado,idTipoPrecio){
             if(Number(rows.EXISTENCIA<=0)){strC='bg-danger text-white'}else{strC='bg-success text-white'};
             let totalexento = 0;
             if (rows.EXENTO==1){totalexento=Number(rows.PRECIO)}
-            str += `<tr id="${rows.CODPROD}">
-            <td>
+            
+            str += `<tr id="${rows.CODPROD}" onclick="getDataMedidaProducto('${rows.CODPROD}','${funciones.quitarCaracteres(rows.DESPROD,'"'," plg",true)}','${rows.CODMEDIDA}',1,${rows.EQUIVALE},${rows.EQUIVALE},${rows.COSTO},${rows.PRECIO},${totalexento},${Number(rows.EXISTENCIA)});" class="border-bottom">
+            <td >
                 ${funciones.quitarCaracteres(rows.DESPROD,'"'," pulg",true)}
                 <br>
                 <small class="text-danger"><b>${rows.CODPROD}</b></small>
@@ -1303,42 +1446,48 @@ async function fcnGuardarNuevoCliente(form){
 };
 
 async function fcnEliminarItem(id){
+    funciones.Confirmacion('¿Está seguro que desea quitar este item?')
+    .then(async(value)=>{
+        if(value==true){
     
-    try {        
-            var data =JSON.stringify({
-                id:id
-            });
-
-            var peticion = new Request('/ventas/tempventas', {
-                method: 'DELETE',
-                headers: new Headers({
-                   'Content-Type': 'application/json'
-                }),
-                body: data
-              });
+            try {        
+                var data =JSON.stringify({
+                    id:id
+                });
+    
+                var peticion = new Request('/ventas/tempventas', {
+                    method: 'DELETE',
+                    headers: new Headers({
+                       'Content-Type': 'application/json'
+                    }),
+                    body: data
+                  });
+            
+                  await fetch(peticion)
+                  
+                  .then(async function(res) {
+                    console.log('Estado: ', res.status);
+                    if (res.status==200)
+                    {
+                        console.log(id.toString());
+                        document.getElementById(id.toString()).remove();
+                        funciones.showToast('item eliminado');
+                        /*await*/ fcnCargarTotal('txtTotalVenta','txtTotalVentaCobro');
+                    }
+                  })
+                  .catch(
+                      ()=>{
+                        funciones.AvisoError('No se pudo remover este producto a la venta actual');
+                      }
+                  )
         
-              await fetch(peticion)
-              
-              .then(async function(res) {
-                console.log('Estado: ', res.status);
-                if (res.status==200)
-                {
-                    console.log(id.toString());
-                    document.getElementById(id.toString()).remove();
-                    //await fcnCargarGridTempVentas('tblGridTempVentas');
-                    funciones.showToast('item eliminado');
-                    await fcnCargarTotal('txtTotalVenta','txtTotalVentaCobro');
-                }
-              })
-              .catch(
-                  ()=>{
-                    funciones.AvisoError('No se pudo remover este producto a la venta actual');
-                  }
-              )
+            } catch (error) {
     
-        } catch (error) {
-
-        }
+            };
+    
+        }        
+    })
+    
 };
 
 async function fcnCargarGridTempVentas(idContenedor){
